llm_provider_mcp/requirements.txt
fastmcp>=0.4.1
google-genai
google-auth
google-auth-oauthlib
openai
python-dotenv
pydantic
